@@ -23,7 +23,7 @@ license = "ISC"
 readme = "README.md"
 repository = "https://github.com/starship/starship"
 # Note: MSRV is only intended as a hint, and only the latest version is officially supported in starship.
-rust-version = "1.60"
+rust-version = "1.64"
 description = """
 The minimal, blazing-fast, and infinitely customizable prompt for any shell! ☄🌌️
 """
@@ -42,44 +42,44 @@ git-repository-max-perf = ["git-features/zlib-ng", "git-repository/fast-sha1"]
 git-repository-faster = ["git-features/zlib-stock", "git-repository/fast-sha1"]
 
 [dependencies]
-chrono = { version = "0.4.22", features = ["clock", "std"] }
-clap = { version = "4.0.18", features = ["derive", "cargo", "unicode"] }
-clap_complete = "4.0.3"
+chrono = { version = "0.4.23", features = ["clock", "std"] }
+clap = { version = "4.0.29", features = ["derive", "cargo", "unicode"] }
+clap_complete = "4.0.6"
 dirs-next = "2.0.0"
 dunce = "1.0.3"
-gethostname = "0.3.0"
-git-features = { version = "0.23.0", optional = true }
+gethostname = "0.4.1"
+git-features = { version = "0.24.1", optional = true }
 # default feature restriction addresses https://github.com/starship/starship/issues/4251
-git-repository = { version = "0.25.0", default-features = false, features = ["max-performance-safe"] }
-indexmap = { version = "1.9.1", features = ["serde"] }
-local_ipaddress = "0.1.3"
+git-repository = { version = "0.29.0", default-features = false, features = ["max-performance-safe"] }
+indexmap = { version = "1.9.2", features = ["serde"] }
 log = { version = "0.4.17", features = ["std"] }
 # nofity-rust is optional (on by default) because the crate doesn't currently build for darwin with nix
 # see: https://github.com/NixOS/nixpkgs/issues/160876
-notify-rust = { version = "4.5.10", optional = true }
+notify-rust = { version = "4.6.0", optional = true }
 nu-ansi-term = "0.46.0"
-once_cell = "1.15.0"
-open = "3.0.3"
+once_cell = "1.16.0"
+open = "3.2.0"
+# update os module config and tests when upgrading os_info
 os_info = "3.5.1"
 path-slash = "0.2.1"
-pest = "2.4.0"
-pest_derive = "2.4.0"
+pest = "2.5.1"
+pest_derive = "2.5.1"
 quick-xml = "0.26.0"
 rand = "0.8.5"
-rayon = "1.5.3"
-regex = "1.6.0"
+rayon = "1.6.0"
+regex = "1.7.0"
 rust-ini = "0.18.0"
 semver = "1.0.14"
-serde = { version = "1.0.147", features = ["derive"] }
-serde_json = "1.0.87"
-sha-1 = "0.10.0"
-shadow-rs = { version = "0.17.0", default-features = false }
+serde = { version = "1.0.149", features = ["derive"] }
+serde_json = "1.0.89"
+sha1 = "0.10.5"
+shadow-rs = { version = "0.18.0", default-features = false }
 # battery is optional (on by default) because the crate doesn't currently build for Termux
 # see: https://github.com/svartalf/rust-battery/issues/33
 starship-battery = { version = "0.7.9", optional = true }
 strsim = "0.10.0"
-systemstat = "=0.2.1"
-terminal_size = "0.2.1"
+systemstat = "=0.2.2"
+terminal_size = "0.2.3"
 toml = { version = "0.5.9", features = ["preserve_order"] }
 toml_edit = "0.15.0"
 unicode-segmentation = "1.10.0"
@@ -104,7 +104,7 @@ features = ["preserve_order", "indexmap"]
 deelevate = "0.2.0"
 
 [target.'cfg(windows)'.dependencies.windows]
-version = "0.42.0"
+version = "0.43.0"
 features = [
   "Win32_Foundation",
   "Win32_UI_Shell",
@@ -114,10 +114,10 @@ features = [
 ]
 
 [target.'cfg(not(windows))'.dependencies]
-nix = { version = "0.25.0", default-features = false, features = ["feature", "fs", "user"] }
+nix = { version = "0.26.1", default-features = false, features = ["feature", "fs", "user"] }
 
 [build-dependencies]
-shadow-rs = { version = "0.17.0", default-features = false }
+shadow-rs = { version = "0.18.0", default-features = false }
 dunce = "1.0.3"
 
 [target.'cfg(windows)'.build-dependencies]
