@@ -101,7 +101,7 @@ pub fn module<'a>(context: &'a Context) -> Option<Module<'a>> {
             String::from(config.truncation_symbol)
         }
     } else {
-        String::from("")
+        String::new()
     };
 
     let path_vec = match &repo.and_then(|r| r.workdir.as_ref()) {
@@ -118,10 +118,10 @@ pub fn module<'a>(context: &'a Context) -> Option<Module<'a>> {
                 let before = before_root_dir(&dir_string, &contracted_path);
                 [prefix + before.as_str(), root.to_string(), after_repo_root]
             } else {
-                ["".to_string(), "".to_string(), prefix + dir_string.as_str()]
+                [String::new(), String::new(), prefix + dir_string.as_str()]
             }
         }
-        _ => ["".to_string(), "".to_string(), prefix + dir_string.as_str()],
+        _ => [String::new(), String::new(), prefix + dir_string.as_str()],
     };
 
     let path_vec = if config.use_os_path_sep {
@@ -137,6 +137,7 @@ pub fn module<'a>(context: &'a Context) -> Option<Module<'a>> {
         config.repo_root_format
     };
     let repo_root_style = config.repo_root_style.unwrap_or(config.style);
+    let before_repo_root_style = config.before_repo_root_style.unwrap_or(config.style);
 
     let parsed = StringFormatter::new(display_format).and_then(|formatter| {
         formatter
@@ -144,6 +145,7 @@ pub fn module<'a>(context: &'a Context) -> Option<Module<'a>> {
                 "style" => Some(Ok(config.style)),
                 "read_only_style" => Some(Ok(config.read_only_style)),
                 "repo_root_style" => Some(Ok(repo_root_style)),
+                "before_repo_root_style" => Some(Ok(before_repo_root_style)),
                 _ => None,
             })
             .map(|variable| match variable {
@@ -183,7 +185,7 @@ fn remove_extended_path_prefix(path: String) -> String {
     }
     // Trim any Windows extended-path prefix from the display path
     if let Some(unc) = try_trim_prefix(&path, r"\\?\UNC\") {
-        return format!(r"\\{}", unc);
+        return format!(r"\\{unc}");
     }
     if let Some(p) = try_trim_prefix(&path, r"\\?\") {
         return p.to_string();
@@ -328,7 +330,7 @@ fn to_fish_style(pwd_dir_length: usize, dir_string: String, truncated_dir_string
         .map(|word| -> String {
             let chars = UnicodeSegmentation::graphemes(word, true).collect::<Vec<&str>>();
             match word {
-                "" => "".to_string(),
+                "" => String::new(),
                 _ if chars.len() <= pwd_dir_length => word.to_string(),
                 _ if word.starts_with('.') => chars[..=pwd_dir_length].join(""),
                 _ => chars[..pwd_dir_length].join(""),
@@ -501,7 +503,7 @@ mod tests {
 
     fn init_repo(path: &Path) -> io::Result<()> {
         create_command("git")?
-            .args(&["init"])
+            .args(["init"])
             .current_dir(path)
             .output()
             .map(|_| ())
@@ -689,7 +691,7 @@ mod tests {
             "{} ",
             Color::Cyan
                 .bold()
-                .paint(convert_path_sep(&format!("/foo/bar/{}/path", strange_sub)))
+                .paint(convert_path_sep(&format!("/foo/bar/{strange_sub}/path")))
         ));
 
         assert_eq!(expected, actual);
@@ -706,7 +708,7 @@ mod tests {
             "{} ",
             Color::Cyan
                 .bold()
-                .paint(convert_path_sep(&format!("~/{}/starship", name)))
+                .paint(convert_path_sep(&format!("~/{name}/starship")))
         ));
 
         assert_eq!(expected, actual);
@@ -724,7 +726,7 @@ mod tests {
             "{} ",
             Color::Cyan
                 .bold()
-                .paint(convert_path_sep(&format!("{}/engine/schematics", name)))
+                .paint(convert_path_sep(&format!("{name}/engine/schematics")))
         ));
 
         assert_eq!(expected, actual);
@@ -788,7 +790,7 @@ mod tests {
             "{} ",
             Color::Cyan
                 .bold()
-                .paint(convert_path_sep(&format!("{}/thrusters/rocket", name)))
+                .paint(convert_path_sep(&format!("{name}/thrusters/rocket")))
         ));
 
         assert_eq!(expected, actual);
@@ -864,7 +866,7 @@ mod tests {
             "{} ",
             Color::Cyan
                 .bold()
-                .paint(convert_path_sep(&format!("{}/rocket", name)))
+                .paint(convert_path_sep(&format!("{name}/rocket")))
         ));
 
         assert_eq!(expected, actual);
@@ -1117,7 +1119,7 @@ mod tests {
         let src_dir = repo_dir.join("src");
         let symlink_dir = tmp_dir.path().join("rocket-controls-symlink");
         let symlink_src_dir = symlink_dir.join("src");
-        fs::create_dir_all(&src_dir)?;
+        fs::create_dir_all(src_dir)?;
         init_repo(&repo_dir).unwrap();
         symlink(&repo_dir, &symlink_dir)?;
 
@@ -1143,7 +1145,7 @@ mod tests {
         let src_dir = repo_dir.join("src/meters/fuel-gauge");
         let symlink_dir = tmp_dir.path().join("rocket-controls-symlink");
         let symlink_src_dir = symlink_dir.join("src/meters/fuel-gauge");
-        fs::create_dir_all(&src_dir)?;
+        fs::create_dir_all(src_dir)?;
         init_repo(&repo_dir).unwrap();
         symlink(&repo_dir, &symlink_dir)?;
 
@@ -1172,7 +1174,7 @@ mod tests {
             .join("above-repo")
             .join("rocket-controls-symlink");
         let symlink_src_dir = symlink_dir.join("src/meters/fuel-gauge");
-        fs::create_dir_all(&src_dir)?;
+        fs::create_dir_all(src_dir)?;
         init_repo(&repo_dir).unwrap();
         symlink(&repo_dir, &symlink_dir)?;
 
@@ -1207,7 +1209,7 @@ mod tests {
             .join("above-repo")
             .join("rocket-controls-symlink");
         let symlink_src_dir = symlink_dir.join("src/meters/fuel-gauge");
-        fs::create_dir_all(&src_dir)?;
+        fs::create_dir_all(src_dir)?;
         init_repo(&repo_dir).unwrap();
         symlink(&repo_dir, &symlink_dir)?;
 
@@ -1244,7 +1246,7 @@ mod tests {
             .join("above-repo")
             .join("rocket-controls-symlink");
         let symlink_src_dir = symlink_dir.join("src/meters/fuel-gauge");
-        fs::create_dir_all(&src_dir)?;
+        fs::create_dir_all(src_dir)?;
         init_repo(&repo_dir).unwrap();
         symlink(&repo_dir, &symlink_dir)?;
 
@@ -1289,7 +1291,7 @@ mod tests {
             .join("above-repo")
             .join("rocket-controls-symlink");
         let symlink_src_dir = symlink_dir.join("src/meters/fuel-gauge");
-        fs::create_dir_all(&src_dir)?;
+        fs::create_dir_all(src_dir)?;
         init_repo(&repo_dir).unwrap();
         symlink(&repo_dir, &symlink_dir)?;
 
@@ -1399,7 +1401,7 @@ mod tests {
             "{} ",
             Color::Cyan
                 .bold()
-                .paint(convert_path_sep(&format!("…/{}/a/subpath", name)))
+                .paint(convert_path_sep(&format!("…/{name}/a/subpath")))
         ));
         assert_eq!(expected, actual);
         tmp_dir.close()
@@ -1424,7 +1426,7 @@ mod tests {
             "{} ",
             Color::Cyan
                 .bold()
-                .paint(convert_path_sep(&format!("~/{}/a/subpath", name)))
+                .paint(convert_path_sep(&format!("~/{name}/a/subpath")))
         ));
         assert_eq!(expected, actual);
         tmp_dir.close()
@@ -1694,12 +1696,14 @@ mod tests {
                 truncation_symbol = "…/"
                 truncate_to_repo = false
                 repo_root_style = "green"
+                before_repo_root_style = "blue"
             })
             .path(dir)
             .collect();
         let expected = Some(format!(
-            "{}{}repo{} ",
-            Color::Cyan.bold().paint(convert_path_sep("…/above/")),
+            "{}{}{}repo{} ",
+            Color::Blue.prefix(),
+            convert_path_sep("…/above/"),
             Color::Green.prefix(),
             Color::Cyan.bold().paint(convert_path_sep("/src/sub/path"))
         ));
@@ -1794,7 +1798,7 @@ mod tests {
             .collect();
         let expected = Some(format!(
             "{} ",
-            Color::Cyan.bold().paint(format!("~/{}/starship", name))
+            Color::Cyan.bold().paint(format!("~/{name}/starship"))
         ));
 
         assert_eq!(expected, actual);
