@@ -79,7 +79,7 @@ pub fn module<'a>(context: &'a Context) -> Option<Module<'a>> {
             })
             .flatten()
             .collect::<String>(),
-        _ => "".to_string(),
+        _ => String::new(),
     };
 
     let main_format = match pipestatus_status {
@@ -114,7 +114,7 @@ fn format_exit_code<'a>(
         }
     };
 
-    let hex_status = format!("0x{:X}", exit_code_int);
+    let hex_status = format!("0x{exit_code_int:X}");
 
     let common_meaning = status_common_meaning(exit_code_int);
 
@@ -323,7 +323,7 @@ mod tests {
         for status in &exit_values {
             let expected = Some(format!(
                 "{} ",
-                Color::Red.bold().paint(format!("❌{}", status))
+                Color::Red.bold().paint(format!("❌{status}"))
             ));
             let actual = ModuleRenderer::new("status")
                 .config(toml::toml! {
@@ -345,7 +345,7 @@ mod tests {
         for (exit_value, string_value) in exit_values.iter().zip(string_values) {
             let expected = Some(format!(
                 "{} ",
-                Color::Red.bold().paint(format!("❌{}", string_value))
+                Color::Red.bold().paint(format!("❌{string_value}"))
             ));
             let actual = ModuleRenderer::new("status")
                 .config(toml::toml! {
