@@ -19,6 +19,7 @@ pub struct DirectoryConfig<'a> {
     pub repo_root_format: &'a str,
     pub style: &'a str,
     pub repo_root_style: Option<&'a str>,
+    pub before_repo_root_style: Option<&'a str>,
     pub disabled: bool,
     pub read_only: &'a str,
     pub read_only_style: &'a str,
@@ -36,9 +37,10 @@ impl<'a> Default for DirectoryConfig<'a> {
             use_logical_path: true,
             substitutions: IndexMap::new(),
             format: "[$path]($style)[$read_only]($read_only_style) ",
-            repo_root_format: "[$before_root_path]($style)[$repo_root]($repo_root_style)[$path]($style)[$read_only]($read_only_style) ",
+            repo_root_format: "[$before_root_path]($before_repo_root_style)[$repo_root]($repo_root_style)[$path]($style)[$read_only]($read_only_style) ",
             style: "cyan bold",
             repo_root_style: None,
+            before_repo_root_style: None,
             disabled: false,
             read_only: "🔒",
             read_only_style: "red",
