@@ -60,6 +60,7 @@ pub const PROMPT_ORDER: &[&str] = &[
     "golang",
     "gradle",
     "haskell",
+    "haxe",
     "helm",
     "java",
     "julia",
@@ -87,6 +88,7 @@ pub const PROMPT_ORDER: &[&str] = &[
     "zig",
     // ↑ Toolchain version modules ↑
     "buf",
+    "guix_shell",
     "nix_shell",
     "conda",
     "meson",
@@ -108,6 +110,7 @@ pub const PROMPT_ORDER: &[&str] = &[
     "time",
     "status",
     "container",
+    "os",
     "shell",
     "character",
 ];
@@ -118,7 +121,7 @@ impl Default for StarshipRootConfig {
         Self {
             schema: "https://starship.rs/config-schema.json".to_string(),
             format: "$all".to_string(),
-            right_format: "".to_string(),
+            right_format: String::new(),
             continuation_prompt: "[∙](bright-black) ".to_string(),
             scan_timeout: 30,
             command_timeout: 500,
