@@ -4,14 +4,27 @@ use crate::config::ModuleConfig;
 use crate::configs::localip::LocalipConfig;
 use crate::formatter::StringFormatter;
 
+use std::io::Error;
+use std::net::UdpSocket;
+
+fn get_local_ipv4() -> Result<String, Error> {
+    let socket = UdpSocket::bind("0.0.0.0:0")?;
+    socket.connect("192.0.2.0:80")?;
+
+    let addr = socket.local_addr()?;
+
+    Ok(addr.ip().to_string())
+}
+
 /// Creates a module with the ipv4 address of the local machine.
 ///
-/// The `local_ipaddress` crate is used to determine the local IP address of your machine.
-/// An accurate and fast way, especially if there are multiple IP addresses available,
-/// is to connect a UDP socket and then reading its local endpoint.
+/// The IP address is gathered from the local endpoint of an UDP socket
+/// connected to a reserved IPv4 remote address, which is an accurate and fast
+/// way, especially if there are multiple IP addresses available.
+/// There should be no actual packets send over the wire.
 ///
 /// Will display the ip if all of the following criteria are met:
-///     - localip.disabled is false
+///     - `localip.disabled` is false
 ///     - `localip.ssh_only` is false OR the user is currently connected as an SSH session (`$SSH_CONNECTION`)
 pub fn module<'a>(context: &'a Context) -> Option<Module<'a>> {
     let mut module = context.new_module("localip");
@@ -28,11 +41,18 @@ pub fn module<'a>(context: &'a Context) -> Option<Module<'a>> {
         return None;
     }
 
-    let localip = local_ipaddress::get().unwrap_or_default();
-    if localip.is_empty() {
-        log::warn!("unable to determine local ipv4 address");
-        return None;
-    }
+    let localip = match get_local_ipv4() {
+        Ok(ip) => ip,
+        Err(e) => {
+            // ErrorKind::NetworkUnreachable is unstable
+            if cfg!(target_os = "linux") && e.raw_os_error() == Some(101) {
+                "NetworkUnreachable".to_string()
+            } else {
+                log::warn!("unable to determine local ipv4 address: {e}");
+                return None;
+            }
+        }
+    };
 
     let parsed = StringFormatter::new(config.format).and_then(|formatter| {
         formatter
@@ -60,19 +80,22 @@ pub fn module<'a>(context: &'a Context) -> Option<Module<'a>> {
 
 #[cfg(test)]
 mod tests {
+    use crate::modules::localip::get_local_ipv4;
     use crate::test::ModuleRenderer;
     use nu_ansi_term::{Color, Style};
 
     macro_rules! get_localip {
         () => {
-            if let Some(localip) = local_ipaddress::get() {
-                localip
-            } else {
-                println!(
-                    "localip was not tested because socket connection failed! \
-                     This could be caused by an unconventional network setup."
-                );
-                return;
+            match get_local_ipv4() {
+                Ok(ip) => ip,
+                Err(e) => {
+                    println!(
+                        "localip was not tested because socket connection failed! \
+                        This could be caused by an unconventional network setup. \
+                        Error: {e}"
+                    );
+                    return;
+                }
             }
         };
     }
